Breaking Change - Added support for Multiple Logins
Added support for multiple logins and support to use it with existing infrastructure instead of creating separate methods. Added  Mirror #d example as its own class. Added support for VivoxConfig so users can pass in their own configuration. Changed some naming for Properties in EasySession. Added global variable to keep track of the current users Login Name/ User name. Have yet to decide how to incorporate this on the low and high levels of the API or to always pass in username in methods